Large refactor.
