+++ src/header.tsx/header.module.css
@@ -1,0 +1,2 @@
+.header {
+}
