+++ src/loading/fallback.tsx
@@ -1,14 +1,13 @@
-import UseNavigationButton from '../navigate/provider';
 import styles from '../../app/page.module.css';
+import { useRouter } from 'next/navigation';
+import LoadingNavigation from './navigation';
 
 export default function LoadingFallback() {
   return (
     <main className={styles.main}>
       <h2>Page Loading</h2>
       <p>The page you are looking for is currently loading. Please wait.</p>
-      <UseNavigationButton route='refresh'></UseNavigationButton>
-      <UseNavigationButton route='back'></UseNavigationButton>
-      <UseNavigationButton route='home'></UseNavigationButton>
+      <LoadingNavigation></LoadingNavigation>
     </main>
   );
 }
