+++ src/loading/provider.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, Suspense } from 'react';
 import LoadingFallback from './fallback';
 
-export default function UseLoading({ children }: { children: ReactNode }) {
+export default function LoadingProvider({ children }: { children: ReactNode }) {
   return <Suspense fallback={LoadingFallback()}>{children}</Suspense>;
 }
