--- src/navigate/types.tsx
@@ -1,3 +1,0 @@
-type Route = 'home' | 'refresh' | 'back';
-
-export type { Route };
